# fix: improve AddLink and HasLink with KeyMatch4 pattern performance