… examples/performance/rbac_with_pattern_large_scale_policy.csv