fix: #781 and fix: #829


```
goos: windows
goarch: amd64
pkg: github.com/casbin/casbin/v2
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
```
|Name       |N        |Time            |
| -------------|:--------------:|--------------:|
| BenchmarkBuildRoleLinksWithPatternLarge-16  (Now)  | 1 | 1380663800 ns/op |
| BenchmarkBuildRoleLinksWithPatternLarge-16  | 1 | 16264016400 ns/op |
| | | |
| BenchmarkBuildRoleLinksWithDomainPatternLarge-16 (Now) | 193 | 6605217 ns/op | 
| BenchmarkBuildRoleLinksWithDomainPatternLarge-16 | 6 | 195383033 ns/op | 
| | | |
| BenchmarkBuildRoleLinksWithPatternAndDomainPatternLarge-16 (Now) | 1| 1828370400 ns/op | 
| BenchmarkBuildRoleLinksWithPatternAndDomainPatternLarge-16 | 1 | 62452898100 ns/op | 
| | | |
| BenchmarkHasLinkWithPatternLarge-16  (Now) | 1120 | 1121191 ns/op | 
| BenchmarkHasLinkWithPatternLarge-16 | 140 | 8890446 ns/op | 
| | | |
| BenchmarkHasLinkWithDomainPatternLarge-16  (Now) | 599990 | 2098 ns/op | 
| BenchmarkHasLinkWithDomainPatternLarge-16 | 10000 | 109491 ns/op | 
| | | |
| BenchmarkHasLinkWithPatternAndDomainPatternLarge-16  (Now) | 967 | 1291647 ns/op | 
| BenchmarkHasLinkWithPatternAndDomainPatternLarge-16 | 62 | 23981366 ns/op | 
| | | |
| BenchmarkRBACModelWithDomainPatternLarge-16 (Now) |  2552 |  467868 ns/op | 
| BenchmarkRBACModelWithDomainPatternLarge-16 |  1200  |  1067164 ns/op |