@closetool 2021-05-25
@Sagilio @hsluoyz 
model.conf
```
[request_definition]
r = sub, obj, act

[policy_definition]
p = sub, obj, act

[role_definition]
g = _, _

[policy_effect]
e = some(where (p.eft == allow))

[matchers]
m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
```

policy.csv
```
p, book_admin, book, get
p, pen_admin, pen, get

g, /user/bob, book_admin
g, /user/*, pen_admin
```

main.go
```
package main

import (
    "fmt"

        "github.com/casbin/casbin/v2"
        defaultrolemanager "github.com/casbin/casbin/v2/rbac/default-role-manager"
        "github.com/casbin/casbin/v2/util"
)

func main() {
    e,err := casbin.NewEnforcer("./model.conf","./policies.csv")
    if err != nil {
        panic(err)
    }
    e.GetRoleManager().(*defaultrolemanager.RoleManager).AddMatchingFunc("",util.KeyMatch)
    _ = e.LoadPolicy()
    fmt.Println(e.Enforce("/user/bob","book","get"))
}
```
Change the placement of `g, /user/bob, book_admin`, `g, /user/*, pen_admin` will lead mistaken result, plz check it.