@closetool 2021-05-25
@Sagilio Please try this
```go
if rm.matchingFunc(namePattern, name1) {
	valueRole, _ := rm.roles.LoadOrStore(key.(string), newRole(key.(string)))
	valueRole.(*Role).addRole(role1)
}
if  rm.matchingFunc(name2, namePattern) {
	role2.addRole(value.(*Role))
}
```